requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
argparse==1.4.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cors==1.0.1
cryptography==45.0.6
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
executing==2.2.0
fastjsonschema==2.21.2
filelock==3.19.1
flask==3.1.2
flask-cors==6.0.1
fonttools==4.59.1
fqdn==1.5.1
future==1.0.0
gevent==25.5.1
gitdb==4.0.12
gitpython==3.1.45
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
ipykernel==6.30.1
ipython==9.4.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
json5==0.12.1
jsonify==0.5
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.8.1
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter-server==2.16.0
jupyter-server-terminals==0.5.3
jupyterlab==4.4.6
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.15
kiwisolver==1.4.9
lark==1.2.2
lxml==6.0.0
markdown-it-py==4.0.0
markupsafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
narwhals==2.1.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.5
notebook-shim==0.2.4
numpy==2.3.2
oauthlib==3.3.1
openai==1.100.2
overrides==7.7.0
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
pdfminer-six==20250506
pdfplumber==0.11.7
pillow==11.3.0
platformdirs==4.3.8
postgrest==1.1.1
prometheus-client==0.22.1
prompt-toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.32.0
psutil==7.0.0
pure-eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydeck==0.9.1
pygments==2.19.2
pyjwt==2.10.1
pymupdf==1.26.3
pyparsing==3.2.3
pypdf2==3.0.1
pypdfium2==4.30.0
pysocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-pptx==1.0.2
pytz==2025.2
pywinpty==3.0.0
pyyaml==6.0.2
pyzmq==27.0.1
realtime==2.7.0
referencing==0.36.2
requests==2.32.5
requests-file==2.1.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.27.0
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.16.1
send2trash==1.8.3
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
storage3==0.12.1
streamlit==1.48.1
strenum==0.4.15
supabase==2.18.1
supabase-auth==2.12.3
supabase-functions==0.10.1
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tldextract==5.3.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250809
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
werkzeug==3.1.3
widgetsnbextension==4.0.14
xlsxwriter==3.2.5
zope-event==5.1.1
zope-interface==7.2
